+++ core/src/main/java/com/arnyminerz/escalaralcoiaicomtat/core/ui/viewmodel/SectorsViewModel.kt
@@ -1,19 +1,19 @@
-package com.arnyminerz.escalaralcoiaicomtat.ui.viewmodel
+package com.arnyminerz.escalaralcoiaicomtat.core.ui.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.liveData
 import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.area.Area
 import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.area.loadAreas
 import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.dataclass.get
-import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.zone.Zone
+import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.sector.Sector
 import com.arnyminerz.escalaralcoiaicomtat.core.shared.AREAS
-import com.arnyminerz.escalaralcoiaicomtat.core.ui.viewmodel.DataClassViewModel
 import timber.log.Timber
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
 
-class ZonesViewModel(private val areaId: String) : DataClassViewModel<Zone>() {
-    override val items: LiveData<List<Zone>> = liveData {
+class SectorsViewModel(private val areaId: String, private val zoneId: String) :
+    DataClassViewModel<Sector>() {
+    override val items: LiveData<List<Sector>> = liveData {
         if (AREAS.isEmpty())
             suspendCoroutine<List<Area>> { cont ->
                 loadAreas(firestore, storage, progressCallback = { current, total ->
@@ -22,9 +22,9 @@ class ZonesViewModel(private val areaId: String) : DataClassViewModel<Zone>() {
                     cont.resume(AREAS)
                 }
             }
-        val zones = AREAS[areaId]?.getChildren(firestore, storage)
-        if (zones != null)
-            emit(zones)
-        else Timber.e("Could not find A/$areaId")
+        val sectors = AREAS[areaId]?.get(zoneId)?.getChildren(firestore, storage)
+        if (sectors != null)
+            emit(sectors)
+        else Timber.e("Could not find Z/$zoneId in A/$areaId")
     }
 }
