+++ core/src/main/java/com/arnyminerz/escalaralcoiaicomtat/core/ui/viewmodel/ZonesViewModel.kt
@@ -1,20 +1,18 @@
-package com.arnyminerz.escalaralcoiaicomtat.ui.viewmodel
+package com.arnyminerz.escalaralcoiaicomtat.core.ui.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.liveData
 import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.area.Area
 import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.area.loadAreas
 import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.dataclass.get
-import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.sector.Sector
+import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.zone.Zone
 import com.arnyminerz.escalaralcoiaicomtat.core.shared.AREAS
-import com.arnyminerz.escalaralcoiaicomtat.core.ui.viewmodel.DataClassViewModel
 import timber.log.Timber
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
 
-class SectorsViewModel(private val areaId: String, private val zoneId: String) :
-    DataClassViewModel<Sector>() {
-    override val items: LiveData<List<Sector>> = liveData {
+class ZonesViewModel(private val areaId: String) : DataClassViewModel<Zone>() {
+    override val items: LiveData<List<Zone>> = liveData {
         if (AREAS.isEmpty())
             suspendCoroutine<List<Area>> { cont ->
                 loadAreas(firestore, storage, progressCallback = { current, total ->
@@ -23,9 +21,9 @@ class SectorsViewModel(private val areaId: String, private val zoneId: String) :
                     cont.resume(AREAS)
                 }
             }
-        val sectors = AREAS[areaId]?.get(zoneId)?.getChildren(firestore, storage)
-        if (sectors != null)
-            emit(sectors)
-        else Timber.e("Could not find Z/$zoneId in A/$areaId")
+        val zones = AREAS[areaId]?.getChildren(firestore, storage)
+        if (zones != null)
+            emit(zones)
+        else Timber.e("Could not find A/$areaId")
     }
 }
